Introduce ENUM_CONSTANT_PARAMETER to operator compiler.
This enables operator methods to declare enum constant parameters.

This feature is disabled in default. To enable this, please insert the following in your build.gradle.

```
asakusafw.javac.processorOption 'com.asakusafw.operator.ENUM_CONSTANT_PARAMETER', 'true'
```

Note that, this feature may not be available for legacy implementations of DSL compiler.